🔵 Mv to resources/table
